requirements.txt
# Core testing dependencies
pytest
pytest-bdd
selenium
pyyaml
pytest-html
browserstack-sdk
